Convert speech web pages to text transcripts
Replace each HTML page with a text file (now with ".txt" extension)
that contains just the "President: <NAME>" line at the top and the
transcript of the speech.  There are some basic HTML formatting
elements still embedded in the speech texts, but they can be dealt
with in a subsequent commit; let's take this one diff at a time.

This commit reflects the output of a truly marvelous Emacs macro,
which unfortunately this margin is too small to contain.