Unduplicate this speech
The speech's text was duplicated, with a few very minor typographical
differences.  I went with the second copy.  This problem exists in the
source at the Miller Center site, so we should report it to them, but
in any case we might as well solve it here right away.

Thanks to Jason Levy for reporting the bug to me.